Take out function to domain layer
