+++ app/src/main/java/org/stepik/android/domain/step_quiz/factory/StepQuizFactory.kt
@@ -1,0 +1,7 @@
+package org.stepik.android.domain.step_quiz.factory
+
+import org.stepic.droid.persistence.model.StepPersistentWrapper
+
+interface StepQuizFactory {
+    fun isStepCanHaveQuiz(stepPersistentWrapper: StepPersistentWrapper): Boolean
+}
