+++ app/src/main/java/org/stepik/android/presentation/step/StepPresenter.kt
@@ -13,15 +13,15 @@ import org.stepik.android.domain.lesson.model.LessonData
 import org.stepik.android.domain.step.interactor.StepInteractor
 import org.stepik.android.domain.step.interactor.StepNavigationInteractor
 import org.stepik.android.domain.step.model.StepNavigationDirection
+import org.stepik.android.domain.step_quiz.factory.StepQuizFactory
 import org.stepik.android.model.comments.DiscussionThread
 import org.stepik.android.presentation.base.PresenterBase
-import org.stepik.android.view.step_quiz.ui.factory.StepQuizFragmentFactory
 import javax.inject.Inject
 
 class StepPresenter
 @Inject
 constructor(
-    stepQuizFragmentFactory: StepQuizFragmentFactory,
+    private val stepQuizFactory: StepQuizFactory,
     private val stepInteractor: StepInteractor,
     private val stepNavigationInteractor: StepNavigationInteractor,
 
@@ -51,11 +51,6 @@ constructor(
         }
 
     private val stepUpdatesDisposable = CompositeDisposable()
-
-    private val isStepHasQuiz: Boolean by lazy {
-        val stepWrapper = stepWrapperRxRelay.value ?: throw IllegalStateException("Cannot be null")
-        stepQuizFragmentFactory.isStepCanHaveQuiz(stepWrapper)
-    }
 
     init {
         compositeDisposable += stepUpdatesDisposable
@@ -118,11 +113,12 @@ constructor(
             .observeOn(mainScheduler)
             .subscribeBy(
                 onNext = { stepWrapper ->
-                    if (!isStepHasQuiz) return@subscribeBy
                     val oldState = this.state
                     if (oldState is StepView.State.Loaded &&
                         oldState.stepWrapper.step.block != stepWrapper.step.block) {
-                        view?.showQuizReloadMessage()
+                        if (stepQuizFactory.isStepCanHaveQuiz(stepWrapper)) {
+                            view?.showQuizReloadMessage()
+                        }
                         this.state = oldState.copy(stepWrapper)
                     }
                 }
