+++ app/src/main/java/org/stepik/android/view/injection/step_quiz/StepQuizModule.kt
@@ -2,6 +2,8 @@ package org.stepik.android.view.injection.step_quiz
 
 import dagger.Binds
 import dagger.Module
+import org.stepik.android.domain.step_quiz.factory.StepQuizFactory
+import org.stepik.android.domain.step_quiz.factory.StepQuizFactoryImpl
 import org.stepik.android.view.step_quiz.ui.factory.StepQuizFragmentFactory
 import org.stepik.android.view.step_quiz.ui.factory.StepQuizFragmentFactoryImpl
 
@@ -11,4 +13,9 @@ abstract class StepQuizModule {
     internal abstract fun bindStepQuizFragmentFactoryImpl(
         stepQuizFragmentFactoryImpl: StepQuizFragmentFactoryImpl
     ): StepQuizFragmentFactory
+
+    @Binds
+    internal abstract fun bindStepQuizFactoryImpl(
+        stepQuizFactoryImpl: StepQuizFactoryImpl
+    ): StepQuizFactory
 }
