+++ app/src/main/java/org/stepik/android/view/step_quiz/ui/factory/StepQuizFragmentFactory.kt
@@ -5,5 +5,4 @@ import org.stepic.droid.persistence.model.StepPersistentWrapper
 
 interface StepQuizFragmentFactory {
     fun createStepQuizFragment(stepPersistentWrapper: StepPersistentWrapper): Fragment
-    fun isStepCanHaveQuiz(stepPersistentWrapper: StepPersistentWrapper): Boolean
 }
